requirements.txt
bcrypt
argon2-cffi
tqdm